Security Fix - ID Validation to Prevent Unauthorized Data Access 🛡️🔒

+++ admin/dBconn/api.php
@@ -106,22 +106,32 @@ function deleteLink()
     $db = $database->connect();
 
     $linkID = $_GET['linkID'];
+    $uno = $_GET['uno'];
+    // Check if the uniqueNo and linkID combination is valid
+    $checkValidity = "SELECT * FROM links WHERE uniqueNo = '$uno' AND linkID = $linkID";
+    $CheckResult = mysqli_query($db, $checkValidity);
+    if (mysqli_num_rows($CheckResult) > 0) {
+        $sql = "DELETE FROM links WHERE linkID='" . $linkID . "'";
+        $result = mysqli_query($db, $sql);
 
-    $sql = "DELETE FROM links WHERE linkID='" . $linkID . "'";
+        if ($result == 1) {
+            echo json_encode(
+                array('message' => 'Link Deleted Successfully')
+            );
+        } else {
+            echo json_encode(
+                array('message' => 'Internal Server Error. Try Again')
+            );
 
-    $result = mysqli_query($db, $sql);
-    echo $result;
-    if ($result == 1) {
-        echo json_encode(
-            array('message' => 'Link Deleted Successfully')
-        );
+        }
     } else {
         echo json_encode(
-            array('message' => 'Internal Server Error. Try Again')
+            array('message' => 'Unauthorized Access')
         );
     }
 }
-;
+
+
 
 function shorty()
 {
@@ -132,12 +142,10 @@ function shorty()
     $uniqueNo = $_POST["uniqueNo"];
     $originalLink = $_POST["originalLink"];
     $shortenLink = $_GET["shortenLink"];
-    // $extra =  $_GET["extra"];
 
     $sql = "INSERT INTO `links` (`uniqueNo`,`originalLink`, `shortenLink`) VALUES ('shorty','$originalLink', '$shortenLink')";
-    // print_r($sql);
+
     $result = $db->query($sql);
-    // var_dump($result);
     if ($result) {
         echo json_encode(
             array('message' => 'Form has been submitted')
